requirements.txt
communex>=0.1.32
PyYAML>=6.0.0
